+++ scripts/systems.js
@@ -8,6 +8,7 @@ import d35e from "./systems/d35e.js";
 import swade from "./systems/swade.js";
 import tormenta20 from "./systems/tormenta20.js";
 import wfrp4e from "./systems/wfrp4e.js"
+import splittermond from "./systems/splittermond.js"
 // ↑ IMPORT SYSTEMS HERE ↑
 
 /**
@@ -26,6 +27,7 @@ export const SYSTEMS = {
             swade,
             tormenta20,
             wfrp4e,
+            splittermond,
             // ↑ ADD SYSTEMS HERE ↑
         }?.[game.system.id];
     }