+++ changelog.md
@@ -5,7 +5,8 @@
 - Fixed issue relating to some systems not generating a new ID for items, which caused false-positives when trying to find similar items on actors that were the source of said items
 - Fixed issue where systems would override core functions on items that modify names and other data, Item Piles will now always call the system's Item specific functions 
   - Fixes issue with PF1 items sometimes showing up as identified when they were unidentified
-- Added support for Warhammer Fantasy Roleplay 4th Ed system 
+- Added support for the Warhammer Fantasy Roleplay 4th Ed system 
+- Added support for the Splittermond system 
 
 ## Version 1.4.5